Address review: shared session auth, required syncTargetId, revert server origin changes
