+++ packages/app-cli/app/command-sync.ts
@@ -15,6 +15,7 @@ const md5 = require('md5');
 import * as locker from 'proper-lockfile';
 import { pathExists, writeFile } from 'fs-extra';
 import { checkIfLoginWasSuccessful, generateApplicationConfirmUrl } from '@joplin/lib/services/joplinCloudUtils';
+import SyncTargetJoplinCloud from '@joplin/lib/SyncTargetJoplinCloud';
 import Logger from '@joplin/utils/Logger';
 import { uuidgen } from '@joplin/lib/uuid';
 import ShareService from '@joplin/lib/services/share/ShareService';
@@ -97,7 +98,7 @@ class Command extends BaseCommand {
 			const checkForCredentials = async () => {
 				try {
 					const applicationAuthUrl = `${Setting.value('sync.10.path')}/api/application_auth/${applicationAuthId}`;
-					const response = await checkIfLoginWasSuccessful(applicationAuthUrl);
+					const response = await checkIfLoginWasSuccessful(applicationAuthUrl, SyncTargetJoplinCloud.id());
 					if (response && response.success) {
 						return response;
 					}
