+++ packages/lib/SyncTargetJoplinCloud.ts
@@ -51,17 +51,7 @@ export default class SyncTargetJoplinCloud extends BaseSyncTarget {
 	}
 
 	public async isAuthenticated() {
-		try {
-			const fileApi = await this.fileApi();
-			const api = fileApi.driver().api();
-			const sessionId = await api.sessionId();
-			return !!sessionId;
-		} catch (error) {
-			if (error.code === 403) {
-				return false;
-			}
-			throw error;
-		}
+		return this.isAuthenticatedViaJoplinServerCompatibleSession();
 	}
 
 	public authRouteName() {
