+++ packages/lib/services/joplinCloudUtils.ts
@@ -94,7 +94,7 @@ export const generateApplicationConfirmUrl = async (confirmUrl: string) => {
 // after an error occurs. E.g.: if the function would throw an error while isWaitingResponse
 // was set to true the next time we call the function the value would still be true.
 // The closure function prevents that.
-export const checkIfLoginWasSuccessful = async (applicationsUrl: string, syncTargetId = 10) => {
+export const checkIfLoginWasSuccessful = async (applicationsUrl: string, syncTargetId: number) => {
 	let isWaitingResponse = false;
 	const performLoginRequest = async () => {
 		if (isWaitingResponse) return undefined;
